merges all controllers
- changes rut to guid param
