+++ src/main/java/xyz/wiraharas/service/TitularService.java
@@ -4,6 +4,6 @@ import xyz.wiraharas.entity.Titular;
 
 public interface TitularService {
 
-    public Titular get(String rut);
+    public Titular get(String guid);
 
 }
